Make tests more robust
- Don't require connection to the internet for api tests to run
  (I was just on a plane!). Instead, we use the helper classes
  used in test_gitpuller for making arbitrary temporary git repos
  in test_api as well
- Don't specify hardcoded names when creating our test git repos -
  instead, create truly temporary directories.
- Stop swallowing exceptions that might occur when trying to fetch
  list of tags / branches in a remote. Without this commit, we
  were swallowing the output of the git command itself
