+++ nbgitpuller/pull.py
@@ -86,30 +86,25 @@ class GitPuller(Configurable):
         This checks to make sure the branch we are told to access
         exists in the repo
         """
-        try:
-            heads = subprocess.run(
-                ["git", "ls-remote", "--heads", "--", self.git_url],
-                capture_output=True,
-                text=True,
-                check=True
-            )
-            tags = subprocess.run(
-                ["git", "ls-remote", "--tags", "--", self.git_url],
-                capture_output=True,
-                text=True,
-                check=True
-            )
-            lines = heads.stdout.splitlines() + tags.stdout.splitlines()
-            branches = []
-            for line in lines:
-                _, ref = line.split()
-                refs, heads, branch_name = ref.split("/", 2)
-                branches.append(branch_name)
-            return branch in branches
-        except subprocess.CalledProcessError:
-            m = f"Problem accessing list of branches and/or tags: {self.git_url}"
-            logging.exception(m)
-            raise ValueError(m)
+        heads = subprocess.run(
+            ["git", "ls-remote", "--heads", "--", self.git_url],
+            capture_output=True,
+            text=True,
+            check=True
+        )
+        tags = subprocess.run(
+            ["git", "ls-remote", "--tags", "--", self.git_url],
+            capture_output=True,
+            text=True,
+            check=True
+        )
+        lines = heads.stdout.splitlines() + tags.stdout.splitlines()
+        branches = []
+        for line in lines:
+            _, ref = line.split()
+            refs, heads, branch_name = ref.split("/", 2)
+            branches.append(branch_name)
+        return branch in branches
 
     def resolve_default_branch(self):
         """
